Remove the usage of query and table name because we have migrated to use a new version (schema.graphql)
