+++ indexer/index-manager/src/lib.rs
@@ -1,10 +1,10 @@
-pub mod core;
-pub mod types;
-pub mod core_helper;
-pub mod builder;
-pub mod config;
+pub mod index_manager;
+pub mod index_manager_helper;
 pub mod hasura;
-pub mod store;
 pub mod hasura_helper;
-pub mod manifest;
-pub mod chain_reader_client;
+pub mod store;
+pub mod builder;
+pub mod ipfs;
+pub mod config;
+pub mod chain_reader;
+pub mod types;
